+++ Branch-SDK/src/io/branch/referral/ServerRequestRActionCompleted.java
@@ -87,6 +87,11 @@ class ServerRequestRActionCompleted extends ServerRequest {
     }
 
     @Override
+    public boolean isGAdsParamsRequired() {
+        return true;
+    }
+
+    @Override
     public void handleFailure(int statusCode, String causeMsg) {
         //No implementation on purpose
     }